License text in all script files
Adds the full MIT license text to the top of all script files since the add-on service's documentation for using a `Dockerfile` to deploy an app to Heroku includes instructions to download `profile.d/borealis-pg-init-ssh-tunnel.sh` on its own. This way, anyone that follows those instructions need do nothing extra to comply with the license.