Support new secure tunnel config var
An upcoming change to the add-on service will switch to the `DATABASE_TUNNEL_BPG_CONN_INFO` config var for secure tunnel connection details. The new name eliminates the soon-to-be-outdated assumption that all secure tunnels for the add-on rely on SSH. This commit futureproofs the buildpack to support the new config var.